reset logic for user session updated
