+++ analytics.js
@@ -39,8 +39,6 @@ import {
   DEFAULT_ERROR_REPORT_PROVIDER,
   ERROR_REPORT_PROVIDERS,
   SAMESITE_COOKIE_OPTS,
-  DEFAULT_SESSION_TIMEOUT,
-  MIN_SESSION_TIMEOUT,
 } from "./utils/constants";
 import { integrations } from "./integrations";
 import RudderElementBuilder from "./utils/RudderElementBuilder";
@@ -52,6 +50,7 @@ import ScriptLoader from "./integrations/ScriptLoader";
 import parseLinker from "./utils/linker";
 import CookieConsentFactory from "./cookieConsent/CookieConsentFactory";
 import * as BugsnagLib from "./metrics/error-report/Bugsnag";
+import { Session } from "./session";
 
 const queryDefaults = {
   trait: "ajs_trait_",
@@ -110,7 +109,7 @@ class Analytics {
     this.cookieConsentOptions = {};
     // flag to indicate client integrations` ready status
     this.clientIntegrationsReady = false;
-    this.sessionTimeout = DEFAULT_SESSION_TIMEOUT;
+    this.session = Session;
   }
 
   /**
@@ -160,138 +159,6 @@ class Analytics {
    */
   executeReadyCallback() {
     this.readyCallbacks.forEach((callback) => callback());
-  }
-
-  /**
-   * A function to initialize session information
-   * @param {object} options    load call options
-   */
-  initializeSession(options) {
-    // Fetch session information from storage if any or initialize with an empty object
-    this.sessionInfo = this.storage.getSessionInfo() || {};
-    /**
-     * By default this.autoTrackSession will be true
-     * Cases where this.autoTrackSession will be false:
-     * 1. User explicitly set autoTrackSession load option to false
-     * 2. When user is manually tracking the session
-     *
-     * Depending on the use case, this.autoTrackSession is set to true/false.
-     */
-    this.autoTrackSession = !(
-      (options && options.autoTrackSession === false) ||
-      this.sessionInfo.manuallyTrackSession
-    );
-    /**
-     * Validate "sessionTimeout" input. Should be provided in milliseconds.
-     * Session timeout: By default, a session lasts until there's 30 minutes of inactivity, 
-     * but you can configure this limit using "sessionTimeout" load option
-     */
-    if (options && typeof options.sessionTimeout === 'number'){
-      // In case user provides 0 as the sessionTimeout, auto session tracking will be disabled
-      if (options.sessionTimeout === 0) {
-        logger.warn('[Session]:: Provided sessionTimeout value 0 will disable the auto session tracking feature.');
-        this.autoTrackSession = false;
-      }
-      // In case user provides a setTimeout value greater than 0 but less than 10 seconds SDK will show a warning
-      // and will proceed with it
-      if (
-        options.sessionTimeout > 0 &&
-        options.sessionTimeout < MIN_SESSION_TIMEOUT
-      ) {
-        logger.warn('[Session]:: It is not advised to set "sessionTimeout" less than 10 seconds');
-      }
-      this.sessionTimeout = options.sessionTimeout;
-    }
-    // If auto session tracking is enabled start the session tracking
-    if (this.autoTrackSession) {
-      this.startAutoTrackingSession();
-    } else if (
-      Object.keys(this.sessionInfo).length &&
-      !this.sessionInfo.manuallyTrackSession
-    ) {
-      this.endSession();
-    }
-  }
-
-  /**
-   * A function to validate current session and return true/false depending on that
-   * @param {number} timestamp
-   * @returns boolean
-   */
-  isValidSession(timestamp) {
-    return (
-      Object.keys(this.sessionInfo).length &&
-      typeof this.sessionInfo.expiresAt === 'number' &&
-      timestamp <= this.sessionInfo.expiresAt
-    );
-  }
-
-  /**
-   * A function to check for existing session details and depending on that create a new session.
-   */
-  startAutoTrackingSession() {
-    const timestamp = Date.now();
-    if (!this.isValidSession(timestamp)) {
-      this.sessionInfo.id = timestamp.toString(); // set the current timestamp
-      this.sessionInfo.expiresAt = timestamp + this.sessionTimeout; // set the expiry time of the session
-      this.sessionInfo.sessionStart = false;
-    }
-    this.storage.setSessionInfo(this.sessionInfo);
-  }
-
-  /**
-   * A public method to start a session
-   * @param {string} sessionId     session identifier
-   * @returns
-   */
-  startSession(sessionId) {
-    if (sessionId && typeof sessionId !== 'string') {
-      logger.error(`[Session]:: "sessionId" should be in string format`);
-      return;
-    }
-    if (
-      !this.sessionInfo.manuallyTrackSession ||
-      !Object.keys(this.sessionInfo).length ||
-      !this.sessionInfo.id
-    ) {
-      this.sessionInfo = {
-        id: sessionId || Date.now(),
-        sessionStart: false,
-        manuallyTrackSession: true,
-      };
-    }
-    this.storage.setSessionInfo(this.sessionInfo);
-  }
-
-  /**
-   * A public method to end an ongoing session.
-   */
-  endSession() {
-    this.sessionInfo = {};
-    this.storage.removeSessionInfo();
-  }
-
-  /**
-   * A function to update an ongoing session.
-   */
-  updateSession(timestamp) {
-    if (!this.sessionInfo.manuallyTrackSession)
-      this.sessionInfo.expiresAt = timestamp + this.sessionTimeout; // set the expiry time of the session
-    if (!this.sessionInfo.sessionStart) this.sessionInfo.sessionStart = true;
-    this.storage.setSessionInfo(this.sessionInfo);
-  }
-
-  /**
-   * A function get ongoing sessionId.
-   */
-  getSessionId(timestamp) {
-    if (
-      !this.sessionInfo.manuallyTrackSession &&
-      !this.isValidSession(timestamp)
-    ) {
-      this.startAutoTrackingSession();
-    }
-    return this.sessionInfo.id;
   }
 
   /**
@@ -949,13 +816,16 @@ class Analytics {
         }
       }
       // If auto/manual session tracking is enabled sessionId will be sent in the context
-      if (this.autoTrackSession || this.sessionInfo.manuallyTrackSession) {
+      if (
+        this.session.autoTrackSession ||
+        this.session.sessionInfo.manuallyTrackSession
+      ) {
         try {
           const timestamp = Date.now();
-          rudderElement.message.context.sessionId = this.getSessionId(timestamp);
-          if (!this.sessionInfo.sessionStart)
+          rudderElement.message.context.sessionId = this.session.getSessionId(timestamp);
+          if (!this.session.sessionInfo.sessionStart)
             rudderElement.message.context.sessionStart = true;
-          this.updateSession(timestamp);
+          this.session.update(timestamp);
         } catch (e) {
           handleError(e);
         }
@@ -1130,7 +1000,14 @@ class Analytics {
     this.userTraits = {};
     this.groupId = "";
     this.groupTraits = {};
-    this.sessionInfo = {};
+    const { manuallyTrackSession } = this.session.sessionInfo;
+    this.session.sessionInfo = {};
+    if (this.session.autoTrackSession) {
+      this.session.startAutoTracking();
+    }
+    if (manuallyTrackSession) {
+      this.session.start();
+    }
     this.storage.clear(flag);
   }
 
@@ -1256,7 +1133,7 @@ class Analytics {
 
     // Session initialization
     try {
-      this.initializeSession(options);
+      this.session.initialize(options);
     } catch (e) {
       handleError(e);
     }
@@ -1506,6 +1383,22 @@ class Analytics {
     }
 
     return returnObj;
+  }
+
+  /**
+   * A public method to start a session
+   * @param {string} sessionId     session identifier
+   * @returns
+   */
+  startSession(sessionId) {
+    this.session.start(sessionId);
+  }
+
+  /**
+   * A public method to end an ongoing session.
+   */
+  endSession() {
+    this.session.end();
   }
 }
