+++ utils/storage/storage.js
@@ -404,7 +404,6 @@ class Storage {
     this.storage.remove(defaults.user_storage_trait);
     this.storage.remove(defaults.group_storage_key);
     this.storage.remove(defaults.group_storage_trait);
-    this.storage.remove(defaults.session_info);
     if (flag) {
       this.storage.remove(defaults.user_storage_anonymousId);
     }
